chore: move operator profile components to one module
