+++ packages/apps/human-app/frontend/src/modules/operator/profile/components/index.ts
@@ -1,0 +1,2 @@
+export * from './profile-disable-button';
+export * from './profile-enable-button';
