+++ packages/apps/human-app/frontend/src/modules/operator/profile/index.ts
@@ -1,0 +1,1 @@
+export * from './profile.page';
