+++ packages/apps/human-app/frontend/src/shared/components/ui/profile/index.ts
@@ -1,0 +1,1 @@
+export * from './profile-list-item';
